Introduce the Matmul class
aqt_matmul.Matmul wraps around the config and creation of lhs/rhs quantizers. This is mostly a refactoring change, with a small out-of-band method provided for diagnostics of tensors that have been useful for analyzing production networks.

The main benefit of the class is that a user only needs to look at one class API to use AQTp.

PiperOrigin-RevId: 451081852