Fix a bug of incorrect weights materialization in TensorQuantizer.
In cl/469309479, TensorQuantizer was modified so as to have just a single method that returns both scale and inv_scale. However, the part of updating quantized variables should have been updated accordingly too, which was missing. It causes to use wrong quantized variables during inference.

There is a unit test that verifies quantized variables but the config was written to make the scale and inv_scale to be 1.0 and thus the test didn't catch the bug. This CL changes `const_bound_coeff` in the config and so the quantized variable should not be the same as the input value.

PiperOrigin-RevId: 471407477