Make AqtEinsum an nn.Module in order to be compatible with setup().
Flax modules can only be instantiated either in setup() or a method with @nn.compact. This CL will fix the case where AqtEinsum is instantiated in a setup(). The problem is that AqtEinsum is currently a dataclass without setup(). Its submodule AqtDotGeneral is instantiated neither in an @nn.compact decorator. Therefore instantiating AqtDotGeneral will cause a complain.

PiperOrigin-RevId: 599903224