Fix the training jax side effect error.
The core reason of jax leak error on the head is that traceable dynamic jax variables have to go through the input arguments of both vjp fwd and vjp bwd functions.

In this case, the aqt config contains keys used for stochastic rounding, so the config needs to be an input argument of the custom vjp.

PiperOrigin-RevId: 601244039