CALIBRATION PROTOTYPE #2: Extend Calibration as nn.Module, and inject it into config.
This does NOT work, with

exception flax.errors.CallCompactUnboundModuleError

: The StaticRangeCalibration (nn.Module) is initialized in flax_srq_e2e_model.py to get injected, but it should be initialized inside @nn.compact or setup. That means, it requires different routine compared to the current calibration.

We should either:

(1) Change Calibration to nn.Module, and update the calibration initialization logic.
(2) Make cfg to get the config for Calibration (Not calibration object directly), and create actual calibration object inside aqt_flax.py (which is not consistent with the current design policy of config.DotGeneralRaw).

PiperOrigin-RevId: 608820589