Rename `abs_val_mapped_to` to `get_quant_bound`.
The `quant_bound` represents the width that the bound corresponds to in the quantization range, and is what the bound needs to be divided by to create the scale.

This name general enough to reflect the function that it serves for both symmetric and asymmetric quantization.

Unrelatedly removes a docstring due to corrosion by a linter blocking submit.

PiperOrigin-RevId: 675627238